undo nwp int16, as need to rescale first

+++ nowcasting_dataset/data_sources/nwp/nwp_data_source.py
@@ -118,7 +118,9 @@ class NWPDataSource(ZarrDataSource):
         selected_data = selected_data.interpolate()
         selected_data = selected_data.sel(target_time=slice(start_dt, end_dt))
         selected_data = selected_data.rename({"target_time": "time", "variable": "channels"})
-        selected_data.data = selected_data.data.astype(np.int16)
+
+        # TODO issue 61, need to rescale before changing to chaning to int16
+        # selected_data.data = selected_data.data.astype(np.int16)
 
         return selected_data
